Extract "same" assertion to own file
