+++ lib/referee.js
@@ -189,18 +189,7 @@ referee.prepareMessage = function msg(message) {
     return message + (/[.:!?]$/.test(message) ? " " : ": ");
 };
 
-referee.add("same", {
-    assert: function (actual, expected) {
-        return samsam.identical(actual, expected);
-    },
-    refute: function (actual, expected) {
-        return !samsam.identical(actual, expected);
-    },
-    assertMessage: "${customMessage}${actual} expected to be the same object as ${expected}",
-    refuteMessage: "${customMessage}${actual} expected not to be the same object as ${expected}",
-    expectation: "toBe",
-    values: actualAndExpectedMessageValues
-});
+require("./assertions/same")(referee);
 
 // Extract/replace with separate module that does a more detailed
 // visualization of multi-line strings
